requirements.txt
flask==3.0.3
tensorflow==2.16.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
pycryptodome==3.20.0
matplotlib==3.9.0
scikit-learn==1.5.0
pillow==10.3.0
werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
